requirements.txt
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.1.31
exceptiongroup==1.2.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
numpy==2.2.4
pandas==2.2.3
python-dateutil==2.9.0.post0
python-telegram-bot==20.3
pytz==2025.2
six==1.17.0
sniffio==1.3.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
